Added `site:core:dns' module
The public functions are easier to use, and take less arguments, while
the internal function is script-parseable, a little more complex in
use, but obviously contains the engine of this module.

There should be no reason to use `dig' or other utilities in any other
`site:*:*' from this version of this file onwards.

This is how these functions tie into one-another:

    :dns:lookup()
       \\\\
        \\\\_________________dns:lookup()
         \\\_________________dns:cname()
          \\___:dns:fqdn()___dns:fqdn()
           \___:dns:qdn()____dns:qdn()