Generalized `site:DEC:ssh' to `site:core:ssh'
Differentiate between a failed ssh connection (returns 255), and a
successful connection that executed a command on the remote host that
returned a non-zero code.  The only issue is if that program also
returned a 255.

The `-t' and `-T' options for ssh are now handled by the
`site:core:ssh->connect()', based on the nature of the arguments supplied.

Added a user defined array for mapping `p' and `P' to interface names.

RW LDAP server array added for modification operations.

Updated to use `site:core:cpf' functions for output.