Overhauled `site:core:ldap': integrity check
The integrity check function has been rewritten, this time it checks
the modification and creation time of all user objects, and does so by
splitting the entire tree into 26 segments, each corresponding to the
letters of the alphabet.  Where a difference is detected, an ldif diff
is printed to screen.  Finally, the pagination function of ldapsearch
is utilised to ensure that LDAP_SIZELIMIT_EXCEEDED is never triggered.