Updated `etc/site.conf': shflags, git
Shflags is now integrated into site itself, not only for module
functions that happen to use it.  This means site itself can now take
certain flags, such as `-h' for help.

A problem with how git was called - assuming it was in the
development directory.  `git' now replaced with `:core:git()',
which takes care of that.