Redesigned `site:core:ldap.search()' functions
The ldap search now accounts for multi-field records.  The only
downside is that the output now concatenates these using a triple pipe
delimiter, which all callers must now be aware of.

Replaced `::ldap:search()' with `:ldap:search()'; as it turns out the
column command has a rather low limit and breaks easily, and that was
the whole purpose of the initial `:ldap:search()' function.  Removing
it made it completely redundant.