Making room for arbitrary output formats
Work has been done to allow any arbitrary output format to be
implemented.

Cleaned up how site determines if the output should be color or not:

    Replaced HAS_TTY with SITE_IN_COLOR, and allow override so color
    can be enforced despite the terminal having an associated tty or
    not.  This is useful for producing ansi output that is then
    translated into color html and sent as an email.

    The logic for setting SITE_IN_COLOR is outside of this module.

Finally, added a new function for replacing direct calls to netcat for
checking if a particular port is open or not - namely `:net:portscan()'.