Updated `etc/site.conf': user-defined interfaces
Also removed dependency on util, as site.conf should be relatively
self-contained.

Work of forcing any exception to cause site to drop out, even in the
case where the process that threw the exception is a forked child
process.

The shflags display of flags for any given function is fixed.

Set the cache directory permissions to 3770.

Upgraded how core:requires work, so it now supports ALL and ANY
checks, and python/perl have been uppercased to make it obvious that
they're special checks (they check for modules, not executables).

Small cleanup.