New `site:core:unit': efficiency, modularity, interface
You can now test individual modules, and not have to test everything
at once.

The auxilliary functions are no longer public, and `all' has been
renamed to `test'.

Only regenerate the script if there is a need to, speeding up the
unit-testing processes which is now getting quite lengthy.

Other than checking that the output of a test is 32-characters long,
also see if the character contents match - before assuming it is an
md5summed output.

... lib/core/unit

... lib/core/unit