New `site:core:remote': PKA `ssh', docstring, shflags
This module has now absorbed various functions from other auxilliary
modules; namely `mon', and `cf' - which have both now disappeared.

This is the migration path:

    mon:remex -> remex->mon

    cf:run -> remex->cfrun

Amended all <tldid> functions to use the shflags global option
instead, and removed all validation checks for it as this is already
done by this stage.

Commented out verbose warnings and added a module docstring

All shflags options for the module that used to be capital-letter have
been lower-cased to align with the new shflags short-option naming
convention.