Updated `module/dns': updated `:dns:fqdn()'
This function now tries to determine if the hint supplied is already
a fqdn (by means of known fqdns and `:dns:qialified()'), and if not,
then do it will fall back to the existing logic/code.

This should simplify things for most callers that had to do this work
themselves, and speed things up in the case that the hint supplied was
already an fqdn.