Updated `module/dns': resolver, wildcard tldid, ut
The resolver search was non-exhaustive: The search for the matching
host resolution left out one case where the `hnh' is the `shn', but
the associated `fqdn' is `shn.dn'; i.e., there is no `usdn' in the
middle.  This has been fixed

Retrospective changes for the new wildcard tldid implemented; see
nima/site#33.

Allow for tldid strings, as opposed to single-character tldids.

Updated unit tests to match changes.