Updated `libsite.sh': new, wildcard tldid, cleanup
Implemented the `not implemented' return code code, so it is now
communicated back to the user.

User defined variables now now first declared in the library, so that
if the user doesn't want to declare them, it doesn't cause unexpected
behaviours.

New `core:tld' function to handle resolving supplied tldid to a tld
safely.

Updated user sample/template file so that variables are no longer
declared as `libsite.sh' now does this already.  Also using the safety
belt for using variables we expect exist (${...?}).

Updated the tldid wildcard as per nima/site#33.

Moved XPLM configuration (interpreter versions) into common area for
both bash and make.

The var area is no longer put in `/var/tmp', but in the user's home
directory.

Removed ssh proxy configuration settings, this is now expected to be
handled via the users' ssh configuration.

SSH options are down to the minimum, as the new idea is to leverage
from the site ssh configuration files instead.

Updated examples, adding a new sample ssh configuration for handling
ssh proxy/jump-boxing, and a more complete set of ldap options.

The triple-character delimiter `|||' was troublesome in some edge
cases, and was hence updated to another global delimiter.
`SITE_DELIM' was used up, so we chose `SITE_DELOM'.  It should only
ever be used when `SITE_DELIM' is also used; hopefully 2 will be
enough.

SSH options now kept here (instead of remote), since other modules on
rare occations may need to call ssh directly, and also since the new
ssh config has been introduced, most options in this variable should
be migrated there anyway.

Don't do an extra module import just to see if it works, as this is
too slow a process.  Just go for it, and hope that the modules are in
good shape.

General cleanup and house-keeping.