New `libexec/ssh': rewrite
The existing python/paramiko ssh interface has been rewritten in ruby
with net-ssh and friends.  Seems to be a little more reliable and less
annoying to use.

Also the python init file no longer needs to do any magic as pyenv has
since been introduced and handles all the work.

The environment variable `SUDO' is not used to determine if the
command is to be executed via sudo or not.  The value of this
environment variable is the secret ID with which the vault will be
queried.